refactor: tools folder organization
